Refactoring: Input label.
