+++ src/components/labelInput/index.js
@@ -3,14 +3,7 @@
  */
 require('!style-loader!css-loader!./index.css')
 
-import toml from 'toml'
-import { tomlString } from '../../utils'
-// import packageJson from '../../../package.json'
-
-import * as alertDialog from '../../uis/alertDialog'
-
-// LocalStorage key to save label data.
-const LSKEY_LABEL_LIST = 'pdfanno-label-list'
+import * as behavior from './behavior'
 
 let $inputLabel
 window.addEventListener('DOMContentLoaded', () => {
@@ -23,8 +16,6 @@ let currentUUID
 
 let _getSelectedAnnotations
 let _saveAnnotationText
-let _createSpanAnnotation
-let _createRelAnnotation
 
 export function setup ({
     getSelectedAnnotations,
@@ -34,215 +25,30 @@ export function setup ({
 }) {
     _getSelectedAnnotations = getSelectedAnnotations
     _saveAnnotationText = saveAnnotationText
-    _createSpanAnnotation = createSpanAnnotation
-    _createRelAnnotation = createRelAnnotation
 
     // Start to listen window events.
     listenWindowEvents()
 
     // Set add button behavior.
-    setupLabelAddButton()
+    behavior.setupLabelAddButton()
 
     // Set trash button behavior.
-    setupLabelTrashButton()
+    behavior.setupLabelTrashButton()
 
     // Set the action when a label is clicked.
-    setupLabelText()
+    behavior.setupLabelText(createSpanAnnotation, createRelAnnotation)
 
     // Set tab behavior.
-    seupTabClick()
+    behavior.seupTabClick()
 
     // Set import/export link behavior.
-    setupImportExportLink()
+    behavior.setupImportExportLink()
 }
 
-// The tab name active.
-let currentTab = 'span'
-
-// Setup the action when a tab is clicked.
-function seupTabClick () {
-    $('.js-label-tab').on('click', e => {
-        const type = $(e.currentTarget).data('type')
-        let d = getLabelListData()
-        const labelObject = d[type] || {}
-        let labels
-        if (labelObject.labels === undefined) {
-            labels = [(type === 'span' ? 'span1' : 'relation1')]
-        } else {
-            labels = labelObject.labels
-        }
-
-        labelObject.labels = labels
-        d[type] = labelObject
-        saveLabelListData(d)
-
-        currentTab = type
-
-        let $ul = $(`<ul class="tab-pane active label-list" data-type="${type}"/>`)
-        labels.forEach((label, index) => {
-            $ul.append(`
-                <li>
-                    <div class="label-list__btn js-label-trash" data-index="${index}"><i class="fa fa-trash-o fa-2x"></i></div>
-                    <div class="label-list__text js-label">${label}</div>
-                </li>
-            `)
-        })
-        $ul.append(`
-            <li>
-                <div class="label-list__btn js-add-label-button"><i class="fa fa-plus fa-2x"></i></div>
-                <input type="text" class="label-list__input">
-            </li>
-        `)
-        $('.js-label-tab-content').html($ul)
-    })
-
-    // Setup the initial tab content.
-    $('.js-label-tab[data-type="span"]').click()
-}
-
-function setupLabelAddButton () {
-    $('.js-label-tab-content').on('click', '.js-add-label-button', e => {
-        let $this = $(e.currentTarget)
-        let text = $this.parent().find('input').val().trim()
-        let type = $this.parents('[data-type]').data('type')
-
-        if (!text) {
-            text = '&nbsp;'
-        }
-
-        let d = getLabelListData()
-        let labelObject = d[type] || { labels : [] }
-        labelObject.labels.push(text)
-        d[type] = labelObject
-        saveLabelListData(d)
-
-        // Re-render.
-        $(`.js-label-tab[data-type="${currentTab}"]`).click()
-    })
-}
-
-function setupLabelTrashButton () {
-    $('.js-label-tab-content').on('click', '.js-label-trash', e => {
-        const $this = $(e.currentTarget)
-        const idx = $this.data('index')
-        const type = $this.parents('[data-type]').data('type')
-
-        let d = getLabelListData()
-        let labelObject = d[type] || { labels : [] }
-        labelObject.labels = labelObject.labels.slice(0, idx).concat(labelObject.labels.slice(idx + 1, labelObject.labels.length))
-        d[type] = labelObject
-        saveLabelListData(d)
-
-        // Re-render.
-        $(`.js-label-tab[data-type="${currentTab}"]`).click()
-    })
-}
-
-function setupLabelText () {
-    $('.js-label-tab-content').on('click', '.js-label', e => {
-        let $this = $(e.currentTarget)
-        let text = $this.text().trim().replace(/&nbsp;/g, '')
-        let type = $this.parents('[data-type]').data('type')
-
-        if (text === '<Empty Label>') {
-            text = ''
-        }
-
-        if (type === 'span') {
-            _createSpanAnnotation({ text })
-        } else if (type === 'one-way' || type === 'two-way' || type === 'link') {
-            _createRelAnnotation({ type, text })
-        }
-    })
-}
-
-function getLabelListData () {
-    return JSON.parse(localStorage.getItem(LSKEY_LABEL_LIST) || '{}')
-}
-
-function saveLabelListData (data) {
-    localStorage.setItem(LSKEY_LABEL_LIST, JSON.stringify(data))
-}
-
-function setupImportExportLink () {
-    $('.js-export-label').on('click', () => {
-        let data = getLabelListData()
-
-        // Transform '&nbsp;' to white space.
-        Object.keys(data).forEach(key => {
-            let labelObject = data[key]
-            let labels = (labelObject.labels || []).map(label => {
-                if (label === '&nbsp;') {
-                    label = ''
-                }
-                return label
-            })
-            labelObject.labels = labels
-        })
-
-        // Conver to TOML style.
-        const toml = tomlString(data)
-        console.log(toml)
-
-        // Download.
-        let blob = new Blob([toml])
-        let blobURL = window.URL.createObjectURL(blob)
-        let a = document.createElement('a')
-        document.body.appendChild(a) // for firefox working correctly.
-        a.download = 'pdfanno.conf'
-        a.href = blobURL
-        a.click()
-        a.parentNode.removeChild(a)
-    })
-
-    $('.js-import-label').on('click', () => {
-        $('.js-import-file').val(null).click()
-    })
-    $('.js-import-file').on('change', ev => {
-        if (ev.target.files.length === 0) {
-            return
-        }
-
-        const file = ev.target.files[0]
-
-        if (!window.confirm('Are you sure to load labels?')) {
-            return
-        }
-
-        let fileReader = new FileReader()
-        fileReader.onload = event => {
-            const tomlString = event.target.result
-            try {
-                const labelData = toml.parse(tomlString)
-
-                // whitespace to '&nbsp'
-                Object.keys(labelData).forEach(key => {
-                    let labelObject = labelData[key]
-                    let labels = (labelObject.labels || []).map(label => {
-                        if (label === '') {
-                            label = '&nbsp;'
-                        }
-                        return label
-                    })
-                    labelObject.labels = labels
-                })
-
-                saveLabelListData(labelData)
-                // Re-render.
-                $(`.js-label-tab[data-type="${currentTab}"]`).click()
-            } catch (e) {
-                console.log('ERROR:', e)
-                console.log('TOML:\n', tomlString)
-                alertDialog.show({ message : 'ERROR: cannot load the label file.' })
-                return
-            }
-        }
-        fileReader.readAsText(file)
-    })
-}
-
+/**
+ * Enable the Label Input UI.
+ */
 export function enable ({ uuid, text, disable = false, autoFocus = false, blurListener = null }) {
-    console.log('enableInputLabel:', uuid, text)
 
     currentUUID = uuid
 
@@ -279,6 +85,9 @@ export function enable ({ uuid, text, disable = false, autoFocus = false, blurLi
     })
 }
 
+/**
+ * Disable the Label Input UI.
+ */
 export function disable () {
     currentUUID = null
     $inputLabel
@@ -321,6 +130,9 @@ function handleAnnotationDeselected () {
     }
 }
 
+/**
+ * Save the text an user wrote, to the annotation ( specified by uuid ).
+ */
 function saveText (uuid) {
     const text = $inputLabel.val() || ''
     _saveAnnotationText(uuid, text)
@@ -330,37 +142,38 @@ function saveText (uuid) {
  * Set window event listeners.
  */
 function listenWindowEvents () {
-    // enable text input.
+
+    // Enable the text input.
     window.addEventListener('enableTextInput', e => {
         enable(e.detail)
     })
 
-    // disable text input.
+    // Disable the text input.
     window.addEventListener('disappearTextInput', e => {
         disable(e.detail)
     })
 
-    // handle annotation deleted.
+    // The event an annotation was deleted.
     window.addEventListener('annotationDeleted', e => {
         treatAnnotationDeleted(e.detail)
     })
 
-    // handle annotation hoverIn.
+    // The event an annotation was hovered in.
     window.addEventListener('annotationHoverIn', e => {
         handleAnnotationHoverIn(e.detail)
     })
 
-    // handle annotation hoverOut.
+    // The event an annotation was hovered out.
     window.addEventListener('annotationHoverOut', e => {
         handleAnnotationHoverOut(e.detail)
     })
 
-    // handle annotation selected.
+    // The event an annotation was selected.
     window.addEventListener('annotationSelected', e => {
         handleAnnotationSelected(e.detail)
     })
 
-    // handle annotation deselected.
+    // The event an annotation was deselected.
     window.addEventListener('annotationDeselected', () => {
         handleAnnotationDeselected()
     })
